imStacks/Slicer.m: fix bug when initializing label images and default jet colormap is used

+++ matImage/imStacks/Slicer.m
@@ -203,7 +203,7 @@ methods
             this.displayRange  = [0 maxi];
             
             if isempty(this.colorMap)
-                this.colorMap = jet(maxi);
+                this.colorMap = jet(double(maxi));
             end
             colormap([this.bgColor ; this.colorMap]);
         end
